# docs/swagger.yaml
basePath: /
definitions:
  presenters.EntityDTO:
    properties:
      name:
        type: string
    required:
    - name
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is an example API with Swagger using Gin.
  title: Template API with Swagger
  version: "1.0"
paths:
  /entities:
    post:
      consumes:
      - application/json
      description: Create a new entity with the provided details
      parameters:
      - description: Entity details
        in: body
        name: entity
        required: true
        schema:
          $ref: '#/definitions/presenters.EntityDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Create a new entity
      tags:
      - entities
  /entities/{id}:
    delete:
      description: Delete an entity by its ID
      parameters:
      - description: Entity ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
      summary: Delete an entity by ID
      tags:
      - entities
    get:
      description: Retrieve an entity by its ID
      parameters:
      - description: Entity ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenters.EntityDTO'
        "404":
          description: Not Found
      summary: Get an entity by ID
      tags:
      - entities
    put:
      consumes:
      - application/json
      description: Update an existing entity with the provided details
      parameters:
      - description: Entity ID
        in: path
        name: id
        required: true
        type: string
      - description: Entity details
        in: body
        name: entity
        required: true
        schema:
          $ref: '#/definitions/presenters.EntityDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenters.EntityDTO'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Update an existing entity
      tags:
      - entities
  /health:
    get:
      description: Returns the health status of the API
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health Check
swagger: "2.0"